refactor: import-file-parsed logic divided
