+++ src/import-service/src/handlers/import-file-parser.ts
@@ -1,42 +1,13 @@
-import csv from 'csv-parser';
-import S3 from 'aws-sdk/clients/s3';
 import { S3Event } from 'aws-lambda';
+import csv from 'csv-parser';
 
-import { BUCKET } from '../constants';
-
-const copyObject = async (client, key: string) => {
-  console.log(`Copy from ${BUCKET}/${key}`);
-  const params = {
-    Bucket: BUCKET,
-    CopySource: `${BUCKET}/${key}`,
-    Key: key.replace('uploaded', 'parsed')
-  }
-
-  const result = await client.copyObject(params).promise();
-  console.log(`${key.split('/')[1]} was copied to parsed/`);
-  return result;
-}
-
-const deleteObject = async (client, key: string) => {
-  const params = {
-    Bucket: BUCKET,
-    Key: key
-  }
-  const result =  await client.deleteObject(params).promise();
-  console.log(`${key.split('/')[1]} was deleted from uploaded/`);
-  return result;
-}
+import { simpleStorageService } from '../services/s3.service';
 
 export const importFileParser = async (event: S3Event) => {
-  const s3 = new S3({ region: 'eu-west-1'});
 
   for (const record of event.Records) {
     const { key } = record.s3.object;
-
-    const s3Stream = s3.getObject({
-      Bucket: BUCKET,
-      Key: key
-    }).createReadStream();
+    const s3Stream = simpleStorageService.createReadStream(key);
 
     try {
       await new Promise((res,rej) => {
@@ -48,15 +19,15 @@ export const importFileParser = async (event: S3Event) => {
             rej(error);
           })
           .on('end', async () => {
-            await copyObject(s3, key);
-            await deleteObject(s3, key);
+            await simpleStorageService.copyObject(key);
+            await simpleStorageService.deleteObject(key);
             res();
           })
       })
     } catch (error) {
       return {
         statusCode: 500,
-        message: error.message
+        body: error.message
       }
     }
   }
